Performing initial cleanup of data
member_id is NULL for many rows. To generate this, used SHA2 algorithm using a set of column values. 
Also, on getting count number of loans taken by members, found that there is a member whose count is 33. On analyzing the data, found that member_id generated for these rows using SHA2 with a set of columns had all NULL values for the columns. This is a outlier.

From the raw data, below data is extracted:
customers
loans
loans repayments
loan defaulters

Columns considered in raw data can be seen in the notebook.